IB/hfi: Add test script support for testing capture of bypass packets.
Updates LoadMoudle.py script to be able to handle passing different module
parameters on different nodes. This is required since bypass packets can only be
sent when capture is off but they can only be landed when capture is on. So each
host needs to behave differently.

INTERNAL: PR 124405 - Support snoop/capture of bypass packets in WFR driver

Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Change-Id: I4fd6eaa6e5ce9caf0d5b5f10a79c8becc466172e