IB/hfi: Fix corrupted egress packet captures
This fix simplifies the walking of the scatter gather list by the
snoop/capture write handler. Code inspection of the update_sge() function shows
that it does not actually operate on pointers to the individual sges but instead
it will copy the sges from the array into the sge state structure. This means
snoop/capture can use that function to update the sge and only need to make a
copy of the sge state structure. This simplification removes the buggy
complicated sge copy on walk loop.

INTERNAL: PR 124840 -  iba_packet_capture tool is providing wrong data during
INTERNAL: SM sweeps

Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Change-Id: Iefb0dd7de6cba63531a2c7cac8504875b6056d8f