IB/hfi: implement STL 'errorinfo' PMA MAD
STL v1 specifies a PM MAD for getting, and resetting the
'errorinfo' for a port. 'errorinfo' is metadata that's
collected on the initial occurrence of certain types of error.
The 'get' request returns all errorinfo, and the 'set'
request resets so that new errorinfo is collected the next
time any of the associated errors happens.

Reviewed-by: Ira Weiny <ira.weiny@intel.com>
Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Arthur Kepner <arthur.kepner@intel.com>

Change-Id: Ic154f4be9919e4fe82a9890d9b08f0d98a31edc8