IB/hfi: Add ring dump logic to debugfs and halt
This patch adds additional ring dump logic to
both the sdes debugfs interface as well as the
sdma engine error handling as a diagnostic tool.

This patch also adds an index parameter to the
descriptor trace to better coorelate the ring entry in a
dump with a a descriptor trace.

INTERNAL: PR 124988 - openmpi verbs- "psm" messages appear while
INTERNAL: running benchmarks

Reviewed-by: Dean Luick <dean.luick@intel.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>