IB/hfi: correct initialization of SendDmaMemory csr
This code was loading a constant value which wasn't
correct for multiple SDMA engines.

That flaw was causing engines that were otherwise idle to
engine to halt with an SDmaPacketTrackingUncErr error.

This patch calculates a uniform per sde credit value
based on define table values and initializes the cnt field
with that and sets the index to the per credit value times the sde index.

INTERNAL: PR 124988 - openmpi verbs- "psm" messages appear while
INTERNAL: running benchmarks

Reviewed-by: Arthur Kepner <arthur.kepner@intel.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>