nd: ndctl class device, and nd bus attributes
This is the position (device topology) independent method to find all
the NFIT-defined busses in the system.  The expectation is that there
will only ever be one "nd" bus discovered via /sys/class/nd_bus/ndctl0.
However, we allow for the possibility of multiple busses and they will
listed in discovery order as ndctl0...ndctlN.  This character device
hosts the ioctl for passing control messages (as defined by the NFIT
spec).  The "format" and "revision" attributes of this device identify
the format of the messages.  In the event an NFIT is registered with an
unknown/unsupported control message format then the "format" attribute
will not be visible.

As it stands currently, only format-interface-code1 (as documented in
the NFIT specification) busses and messages are supported.

Signed-off-by: Dan Williams <dan.j.williams@intel.com>