nd: dimm (nfit-memory-devices) in sysfs
Register the dimms described in the nfit as devices on a nd_bus, named
"dimmN" where N is a global ida index.  The dimm numbering per-bus will
be contiguous, since we only allow a single nd_bus to be registered at
at time.  However, eventual dimm-hotplug invalidates this property and
dimms should be addressed via NFIT-handle.

Signed-off-by: Dan Williams <dan.j.williams@intel.com>