nd: regions (block-data-window and persistent memory)
A "region" device represents the maximum capacity of a block-data-window,
or an interleaved spa range (direct-access persistent memory), without
regard for aliasing.  Aliasing is resolved by the label data on the dimm
to designate which exclusive interface will access the aliased data.
Enabling for the label-designated sub-device is in a subsequent patch.

Both "region" types are defined in the NFIT System Physical Address (spa)
table by a unique index per "region".  In the case of persistent memory the
spa-range describes the direct memory address range of the storage
(NFIT_SPA_PM).  A block "region" region (NFIT_SPA_DCR) points to a DIMM
Control Region (DCR) space.  That same DCR is referenced by a
block-data-window (BDW) to describe the capacity of the block-accessible
storage.

The name format of "region" devices is "regionN-X" where:
N - ida allocated bus id from the parent ndbus
X - spa table index from NFIT describing the type and form of the region

"region"s have 4 attributes "type", "size", "interleave_ways", and
"mapping"s where:
- type: either "block" or "pm"

- size: the block-data-window accessible capacity or the span of the
  spa-range in the case of pm.

- interleave_ways: the number of dimms contributing to the capacity of
  the spa range in the pm case or "1" in the case of a block region.

- mappingN: a tuple describing a dimm's contribution to the region's
  capacity in the format (<nfit-dimm-handle>,<offset>,<size>).  For a
  pm-region there will be at least one mapping per dimm in the interleave
  set, but possibly more if the interleave algorithm forces discontiguous
  ranges to be consumed from a dimm.  For a block-region there is only
  "mapping0" listing the starting dimm offset of the block-data-window and
  the available capacity of that window (matches "size" above).

For now the number of mappings per "region" is hard coded at eight.  If
that turns out to not be enough a "mappings" attribute clarifies how
many there are supposed to be, but the goal is to set the hard coded
number at the architectural max.

Signed-off-by: Dan Williams <dan.j.williams@intel.com>