nd: regions (block-data-window and persistent memory)
A "region" device represents the maximum capacity of a block-data-window,
or an interleaved spa range (direct-access persistent memory), without
regard for aliasing.  Aliasing is resolved by the label data on the dimm
to designate which exclusive interface will access the aliased data.
Enabling for the label-designated sub-device is in a subsequent patch.

Both "region" types are defined in the NFIT System Physical Address
(spa) table.  In the case of persistent memory the spa-range describes
the direct memory address range of the storage (NFIT_SPA_PM).  A block
"region" region (NFIT_SPA_DCR) points to a DIMM Control Region (DCR)
space.  That same DCR is (optionally) referenced by a block-data-window (BDW)
to describe the access mechanism and capacity of the block-accessible
storage.  If the related BDW is not published then the dimm is only
available for control/configuration commands.

The name format of "region" devices is "regionN" where, like dimms, N is
a global ida index assigned at discovery time.  This id is not reliable
across reboots nor in the presence of hotplug.  Look to attributes of
the region or static id-data of the sub-namespace to generate a
persistent name.

"region"s have 3 generic attributes "size", "interleave_ways", and
"mapping"s where:
- size: the block-data-window accessible capacity or the span of the
  spa-range in the case of pm.

- interleave_ways: the number of dimms contributing to the capacity of
  the spa range in the pm case or "1" in the case of a block region.

- mappingN: a tuple describing a dimm's contribution to the region's
  capacity in the format (<nfit-dimm-handle>,<offset>,<size>).  For a
  pm-region there will be at least one mapping per dimm in the interleave
  set, but possibly more if the interleave algorithm forces discontiguous
  ranges to be consumed from a dimm.  For a block-region there is only
  "mapping0" listing the starting dimm offset of the block-data-window and
  the available capacity of that window (matches "size" above).

The max number of mappings per "region" is hard coded per the contraints of
sysfs attribute groups.  That said the number of mappings per region should
never exceed the maximum number of possible dimms in the system.  If the
current number turns out to not be enough then the "mappings" attribute
clarifies how many there are supposed to be. "32 should be enough for
anybody...".

Signed-off-by: Dan Williams <dan.j.williams@intel.com>