IB/hfi: Fix a lost Eager update flag
The change for supporting multiple packets per eager
buffers (d766b4c6708b124fe464a6fa9e2a08e8ccf169e5)
changed how the flag inidcating whether to update the
eager buffer head was set.

That indroduced a bug that cause the flag, once set,
to be "lost" on subsequent loop iterations. This was
due to an wrong assumption that the header queue head
would be updated in the same loop iteration as the one
when the flag was set.

INTERNAL: PR 124871 - hfi driver to support multi-packets
INTERNAL: per eager buffer

Reviewed-by: Dean Luick <dean.luick@intel.com
Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>
Change-Id: Ie7557a67bdd7b6dd42ed613c44c138ded7f123b7