IB/hfi: move sleep/wakeup qp refcount
The existing implementation precludes the implemention
of a common API for verbs and PSM.

The patch now increments the refcount when the qp
is queued to a wait head and decrements the refcount
when the wakeup occurs.

This fix allows for the DMA progress mechanism to
not have refcount maintenance, while continuing
that as a verbs specific implementation detail.

The patch also uses the qp_wakeup_qp() to replace
the multi-line instances of the same code.

INTERNAL: PR 124933 - Implement SDMA common API

Change-Id: I7801b243eaf2beac6865c8a9040eddc25bf62b70
Reviewed-by: Mitko Haralanov <mitko.haralanov@intel.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>