IB/hfi: implement common SDMA new/old progress routine
The patch implements a common sdma progress routine
that handles old verbs, new verbs (not yet coded), and
new PSM (not yet coded).  The handling of old txreqs
uses the SDMA_TXREQ_F_FREEDESC and SDMA_TXREQ_F_FREEBUF
to allow the same progress routine to distinguish the old
verbs code from the new cases.

The patch requires an API change in the iowait structure
to present two additional callbacks: sleep and wakeup.

sleep is intended for use when the ring is "full" and
the iowait needs to be "queued" in a ULP specific way
to "wait" for a wakeup call.  This probably means
that the busycb that is presented to the sdma_send_*
functions can be eliminated.  The wakeup is used to
trigger progress via the qib_qp_wakeup() routine.

There are bug fixes in sdma_send_txreq() to add
the missing set of the tx ring pointers start_idx
and next_descq_idx and to properly handle the tail
ring wrap case.

To account for lack of coalesce and descriptor
extending code, the NUM_DESC is increased from
4 to 6.

The patch also contains (from Mitko!) the ahg
descriptor builder.

TODO:
- sdma_send_txlist()
- non-dword sized tlen
  * requires a dummy pre-mapped descriptor
    to xmit 1 - 3 bytes
- verbs use of new API
- Coalesce logic for > 64 descriptors
- Descriptor list extending > NUM_DESC

INTERNAL: PR 124933 - Implement SDMA common API

Change-Id: Ic526b9f27e43f21e7ab76d1bce29a302c1989169
Reviewed-by: Mitko Haralanov <mitko.haralanov@intel.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>