IB/hfi: return replay buffer depths in portinfo queries
STL's portinfo SMA query returns the number of LTPs which can be
accomodated in the HFI's replay buffer, and the number of LTPs
which can be in flight (on the wire) to our link neighbor. Until
now we'd been returning 0 for both. Fix that here.

INTERNAL: PR 125002 - WFR PortInfo.ReplayDepth is 0 in SMA query

Reviewed-by: Ira Weiny <ira.weiny@intel.com>
Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Arthur Kepner <arthur.kepner@intel.com>