IB/hfi: return VL preemption-related info in SMA queries
STL specifies a 'preemption matrix' (indicating which VLs can
preempt other VLs), and 'preemption elements' (similar to the
high/low priority weight arrays used in VL arbitration, but
specifically for use in preemption). WFR does not support
preemption, so a 'get' will return the hardcoded power on
defaults (from section 20.2.2.6.15 of the STLv1 spec.), and a
'set' will return a status indicating the method/attribute
combination is unsupported.

Also changed fm_get_table() so that it returns the size of the
table, which makes things a bit tidier.

INTERNAL: PR 124992 - Preemption Matrix not being reported
INTERNAL: properly by SMA

Reviewed-by: Ira Weiny <ira.weiny@intel.com>
Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Arthur Kepner <arthur.kepner@intel.com>

Change-Id: I73f017788408c4a1029ba2a714b433ad549af273