IB/hfi: add SDMA pad logic
A WFR feature is the ability to send from arbitrary
alignments for arbritrary sizes.   SusieQ needed
dword alignment sizing.

One detail is that the pbc size (in dwords) must match
the sum of the size of the descriptors (in bytes).

This means that user data may need to be padded with up
to three bytes of zero filled data with a trailing last
descriptor.

This patch adds that logic to the new common SDMA mid-layer.

This patch was tested with verbs code under development
with a 1 byte payload.  The code correctly produced an
extra descriptor of 3 bytes referenced the pad dma-able
allocation as revealed by the SDMA descriptor trace.

Additionally, the sde default_desc1 field is moved to
the hfi_devdata for two reasons.  First is is constant
for all sdes and two it avoids changing the commmon API
since the tx close routine needs it but doesn't know the
sde.

INTERNAL: PR 124933 - Implement SDMA common API

Change-Id: I58674cb5ce1995f814af2cba28fd4b5a8a095788
Reviewed-by: Mitko Haralanov <mitko.haralanov@intel.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>