IB/hfi: add sdma_send_txlist() call
This patch adds the sdma_send_txlist API call.

A new static inline submit_tx() is used to to the low level submit and
is used by the list based send and the single tx send.

In addition, a BUG_ON() is deleted in favor of an -EINVAL return.  The
appropriate API comments are adjusted.

INTERNAL: PR 124933 - Implement SDMA common API

Change-Id: I9ca11efcb26e5b520877c7c985b2b0b8e21298aa
Reviewed-by: Mitko Haralanov <mitko.haralanov@intel.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>