IB/hfi: Add in sdma idle interrupt handling
The code has a JAG crutch that forces a per packet interrupt.

This patch removes the crutch and enables the idle interrupt
There is an issue with simics in that the idle interrupt
may not work, so the code adjusts automatically to work
for wfr-event >= 46, where it will be fixed.

Additional tracing has also been added to confirm the
idle interrupts are being seen.

The sdma_idle_cnt module parameter is changed from
SusieQ SPC clocks to nano seconds.   The code computes
the correct cclock based on FPGA or ASIC using a constant
ns/cycle. For simics any non-zero value works.

[HSD 291249 documents the simics issue and wfr-event 4.8.46
will have the fix]

INTERNAL: PR 124101 - SDMA support in Driver with Functional Simulator -
INTERNAL: Multiple SDMA Engines

Change-Id: I916aeb397c81c87ece324fb5c3a701af0fc95b56
Reviewed-by: Dean Luick <dean.luick@intel.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>