IB/hfi: Add 14b CRC sideband option
The POR is to automatically enable sideband credit return
with 14b link CRC mode.  This change adds:
o A set or clear (could change on link bounce) of the
  apropriate SendCmCtrl field.
o Add a new module parameter to allow it to be turned off.

In addition:
o Fix a typo in a print.
o Add a print of the final DC CSR LCB mode value.  This,
  like the other prints in this routine, will either be
  removed or converted to a trace in a future clean-up step.

INTERNAL: PR 125194 - Add 14b CRC sideband option

Reviewed-by: Arthur Kepner <arthur.kepner@intel.com>
Signed-off-by: Dean Luick <dean.luick@intel.com>