IB/hfi: add verbs use of new SDMA API
This patch adds in verbs use of the new SDMA API.

The highlights include:
- A rewrite of qib_verbs_send_dma()
- Two new supporting helper routines:
  o build_verbs_tx_desc
  o build_verbs_ulp_payload

The code contains an enhancement vs. qib that allows for DMA from
arbitrary addresses and sizes.

There is dead code removal from verbs and the common code as
as well as eliminating the start_idx in the sdma_txreq that
was made obsolete with this new verbs implementation.

INTERNAL: PR 124934 - Implement verbs use of common SDMA API

Change-Id: I2211bfab0c0870f39c5b87a061f8c1bdea43383f
Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Reviewed-by: Mitko Haralanov <mitko.haralanov@intel.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>