nd_region: support for legacy nvdimms
The NFIT region driver is an intermediary driver that translates NFIT
defined "region"s into "namespace" devices that are consumed by peristent
memory block drivers.  A "namespace" is a sub-division of a region.

Support for NVDIMM labels is reserved for a later patch.  For now,
publish 'nd_namespace_io' devices which are simply memory ranges with no
regard for dimm boundaries, interleave, or aliasing.  This also adds a
"nstype" attribute to the parent region so that userspace can know ahead
of time the type of namespaces a given region will produce.

Signed-off-by: Dan Williams <dan.j.williams@intel.com>