nd_region: initial driver implementation, support for legacy nvdimms
From the Kconfig text for CONFIG_NFIT_REGION:

  The "NFIT Region Driver" is an intermediary driver that
  translates NFIT defined regions into "namespace" devices that
  are consumed by peristent memory block drivers.  A "namespace"
  is a sub-division of a region.

Support for NVDIMM labels is reserved for a later patch.  For now,
publish 'nd_namespace_io' devices which are simply memory ranges with no
regard for dimm boundaries, interleave, or aliasing.

Now that we are loading "nd driver"s against "nd device"s two new
attributes are exposed per device, "devtype" and "modalias" where:
- devtype: a string representing the type of device

- modalias: a module alias that udev or other module-tooling would use
  to search for a driver module.  This is in support of provding
  userspace a method to "enable"/"disable" devices by mapping that to
  "bind"/"unbind" kernel-driver-core operations.

These attributes are identical to DEVTYPE and MODALIAS variables that
are published in the uevent when a region is added.

This also adds a "nstype" attribute to the parent region so that
userspace can know ahead of time the type of namespaces a given region
will produce.

Signed-off-by: Dan Williams <dan.j.williams@intel.com>