IB/hfi: add drain delay/wakeup
This patch extends the iowait capability with a common API
change.

Prior to this patch, the sdma_busy counter was incremented
in the sdma.c senders, but decremented in the verbs completion
callback.   This was oddly asymetric.

This patch moved the wait queue to iowait and provides
new iowait methods to wait on a drain and trigger drain waiters.
qp destroy logic now uses the new entrances.  The sdma progress
routine now automatically triggers the wait queue on the
sdma_busy reaching zero.

The complete routine now has a third argument which is true
if the iowait sdma_busy has reached zero.  This is used
by verbs in error state handling to re-start a qp that
was stalled on non-zero sdma_busy as part of error state
flushing from a send engine.

INTERNAL: PR 124933 - Implement SDMA common API

Change-Id: I5c50c85014b6887eeef2f9f4f54a303a1b5d6900
Reviewed-by: Mitko Haralanov <mitko.haralanov@intel.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>