Initial FXR driver support for Command Queues
Command Queues are assigned in a pair of TX and RX queues.
Currently just a single pair can be assigned per PID, but the
interface allows flexibility if we want to allow many-to-one.
The CQ head (read) indices are currently managed to be in separate
cachelines and memory is allocated at module load time.
The API and token supports the ability for the driver to choose other
placement of the CQ head indices without breaking compatibility, and
can be changed later if we don't want multiple CQ head indices in the
same PAGE.
CQ drain/reset is not yet implemented.

INTERNAL: PR 124888 - Implement prototype FXR driver

Reviewed-by: David Addison <david.addison@intel.com>
Reviewed-by: John Gregor <john.a.gregor@intel.com>
Signed-off-by: Brian Welty <brian.welty@intel.com>
Change-Id: I3817ebc1ec55e6f62cb4917c3683b3c775d2440a