IB/hfi: fix leftover bits in descs for non-first adds
Commit 414d32e1
("IB/hfi: Correct initial desc0 for descriptors after the first")
missed a case where the first bit was in a descriptor other
than the first.

This patch alters make_tx_sdma_desc() to accept the tx
and the routine adds the first bit for the first descriptor
and insures it is zero otherwise.

Adjusts are also made to the pad logic to react to the
new make_tx_sdma_desc() signature.

INTERNAL: PR 124933 - Implement SDMA common API

Change-Id: I0189f1fdea96e21769740c42ca5d06902270e7a3
Reviewed-by: Mitko Haralanov <mitko.haralanov@intel.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>