IB/hfi: Add LCB accessors
Add an interface for LCB access and update the LCB accessors
with the interface.   The accessor interface routines must
coordinate with the current link state and the 8051 for access
to the LCB registers.  This also quiets informational 8051 comand
prints so that the frequent LCB accessess do not flood
/var/log/messages.

Detail hilights:
o Add accessor calls to the SMA - for MAD processing.
o Add accessor calls to the CSR accessor device - for
  hfidiags accesses.
o Remove accessor checks for LCB access in the interrupt
  clear down routine.  We now only allow LCB interrupts when
  we know we have access.
o The accessor calls take the same lock as the link state
  machine - reassigning the access and adjusting the use
  count must operate as a unit against a state change.
o Add a workaround for the simulator leaving the linkup bit set
  in the 8051 info register.

INTERNAL: PR 124799 - Create function in driver to gain access
INTERNAL: to registers that are not owned by host or need
INTERNAL: special handling

Change-Id: I912ced9c8e67bf3bfb245b7d2dd577fdec8991eb
Reviewed-by: Arthur Kepner <arthur.kepner@intel.com>
Signed-off-by: Dean Luick <dean.luick@intel.com>