IB/hfi: check if CCA is initialized before doing CCA queries
If the HFI driver was not configured to use CCA, and the FM tried
to set CCA parameters (due to an inconsistent configuration between
driver, and FM) we'd dereference a NULL pointer and crash. Add a
macro which indicates whether or not the CCA structures have been
allocated, and check this prior to any CCA-related query.

INTERNAL: 125187 - Panic in WFR driver, version 5-127, bit stream
INTERNAL: 0x1a

Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Reviewed-by: Ira Weiny <ira.weiny@intel.com>
Signed-off-by: Arthur Kepner <arthur.kepner@intel.com>

Change-Id: I7f43fa2564065fefb24e165a5a03f72abcd3aa73