IB/hfi: Add STL single lane speed
This adds the 1x lane speed as a recognized speed. This is
specifically to work around the issue of a single lane
for FPGA systems.  This does not address the larger issue
of additional speeds that have no IB enum.

INTERNAL: PR 125352 - Reporting of 1x link in prr firmware
INTERNAL: 10.0.0.0.137 has caused IPoIB to stop working

Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Dean Luick <dean.luick@intel.com>