IB/hfi: Correctly set the context RcvArray starting point
The starting RcvArray entry for each context is
based on the number of RcvArray groups per context and
any left-over groups spread out to the kernel and
user contexts.

Therefore, to properly find the starting point and
prevent contexts from overlapping, the calculation of
the starting point needs to take into account the context
number, the number of groups, and the number of extra
groups for both kernel and user contexts.

INTERNAL: PR 125358 - WFR driver program RecvArray with overlap

Reviewed-by: Dean Luick <dean.luick@intel.com>
Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>
Change-Id: I19cc690870962d289ea831bb27344170ebf73944