IB/hfi: correctly handle 'link_downed' counter in error summary counts
The error count summary which is returned in
STL_PM_ATTRIB_ID_DATA_PORT_COUNTERS PMA queries is a 64 bit quantity,
but the 'link_downed' counter, which is included in the error count
summary is 32 bits. Cast 'link_downed' to a u64 before adding it to
the summary value.

INTERNAL: PR 125258 - iba_pmaquery reports "Error Counter Summary
INTERNAL: 4294967296"

Reviewed-by: Ira Weiny <ira.weiny@intel.com>
Signed-off-by: Arthur Kepner <arthur.kepner@intel.com>