IB/hfi: read csr vs. dma head based on module parameter
The current code will only read the CSR when it SDMA engine is not running.

There are cases where using the DMA head results in a stale
head and an SDMA engine stall.

INTERNAL: PR 125431 - Read CSR is needed as a detour for seeing an
INTERNAL: outdated DMA head during progress processing

Reviewed-by: Mitko Haralanov <mitko.haralanov@intel.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>