IB/hfi: Fix snoop tests to run with regression suite with PRR
This patch adds snoop tests into the defualt test suite. It fixes a number of
issues with the scripts relying on specific LIDs being assigned. It also stream
lines how the tests run and removes most module reloads. There are now a total
of 2 required, which is the minimum.

This fix also removes some tests that are better suited as unit tests than
regression tests such as filtering and receiving packets with corrupted DLIDs,
and bypass packets.

INTERNAL: PR 125408 - Snoop tests fail with PRR

Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Change-Id: I5a8a5527a3d3d485f27f9e70df3884b3a49d51ec