IB/hfi: Don't over-allocate send contexts
The driver did not boundry-check the number of
send contexts it was allocating against the total
number available on the chip and as a result was
allocating more than it should.

With this fix we have to adjust what is being
reported to user land through the 'nctxts' sysfs
file. PSM uses this number for its rank/context
allocation and it needs an accurate count of the
number of ranks it can start.

INTERNAL: PR 125222 - "hfi0: PIO Error: PioWriteBadCtxt
INTERNAL: (unhandled)" error and kernel oops while
INTERNAL: running mpi_stress/psm2 with shared contexts

Reviewed-by: Dean Luick <dean.luick@intel.com>
Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>
Change-Id: I4adfe8cadab2edd860238a46fa6c2016423c2d83