IB/hfi: Fix crash in hfi_close
Fix crash caused by attempting to dereference
a NULL ctxtdata pointer. This happened in
cases where the context failed to initialize for
any reason. In such cases, hfi_close would still
be called. However the uctxt pointer would be
NULL as a context was not allocated.

The driver attempted to dereference the pointer
prior to checking it.

INTERNAL: PR 125515 - Kernel backtrace during context
INTERNAL: close when context allocation has failed.

Reviewed-by: Arthur Kepner <arthur.kepner@intel.com>
Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>