IB/hfi: Implement User Send Context Reset command
This change introduces the Send Context Reset
command that can be used to reset a halted
user send context.

The HW does not define a send context behavior
when a process writes to the send context memory
while it is being reset. Therefore, it is the
responsibility of all user software using WFR
send context to guarantee that no process will
be writing to the send context memory during the
reset procedure.

If user software breaks that guarantee and allows
a process to write to the send context memory while
it is being reset, only that particular send context
would be broken. This behavior is similar to the
user process halting the send context in the first
place. Therefore, it is acceptable that the driver
does not impose any quiescence restrictions.

INTERNAL: PR 124594 - Driver should support PSM to
INTERNAL: recover from halted send context

Reviewed-by: Dean Luick <dean.luick@intel.com>
Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>
Change-Id: I6759de717367111e9c25e33d49c9e0688bace218