IB/hfi: Fix user interrupts
Call a handler for user space interrupts.  This is for both
receive urgent interrupts and receive available interrupts.

Implement hdrqempty() which checks for un-processed received packets.
This is used in the user poll mechansim where the user wants to wait
for packets received.

While there: Replace f_hdrqempty() with a direct call.

INTERNAL: PR 125532 - Implement hdrqempty
INTERNAL: PR 125430 - RcvUrgentInt interrupt is not hooked up to the
INTERNAL: user poll mechanism

Change-Id: Ia790acda83be25c6ba94a34083652ec6e68db21d
Reviewed-by: Mitko Haralanov <mitko.haralanov@intel.com>
Signed-off-by: Dean Luick <dean.luick@intel.com>