IB/hfi: Prevent list dmawait list corruption
The SDE's dmawait list is protected by the ibdev's
pending_lock. While that particular lock does not
disable interrupts, it is nested within the general
SDE's lock which does.

Since the dmawait list can be modified from within
interrupt context, the user SDMA code has to properly
protect the list to prevent corruption. Therefore, the
user packet queue is added to the wait list in the
sleep/defer callback since that callback is called with
the engine's lock held and interrupts disabled.

INTERNAL: PR 125583 - SDMA crash system with big job

Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>