IB/hfi: fix ack bit decode in trace
The A bit is not decoded correctly due to an endian issue.

This patch corrects the issue.

It also corrects an issue where a packet without extended
headers was printing leftover data in the trace.

INTERNAL: PR 125648 - ibhdr trace doesn't decode the RC ack bit and
INTERNAL: headers without extensions incorrectly

Change-Id: Id6fa0fbf3aa6b170460d4566af2d1e3b350e9b5c
Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>