IB/hfi: Dump module load address when loading.
It is useful to know the address the driver is loaded at when looking at a stack
trace and trying to pin point the exact line of code causing the issue. This can
be used to feed to gdb when loading symbols to make this easier.

INTERNAL: PR 125496 - Improve driver regression testing

Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Change-Id: Ic3c356cc9cd2a7d88bcdfebf0005fc23c3220081