IB/hfi: initialize pkey_index in CQEs for HFI-local LR SMPs
When a Lid Routed SMP is sent from, and destined to the same HFI (i.e.,
the SMP never travels on 'the wire'), the pkey_index in the resulting
CQE was being set to 0 (by convention is the index of the management
pkeys is either 1, or 2). When the FM is configured to reject SMPs with
'spoofed' pkeys, these SMPs are rejected, and the port cannot be
brought to the ACTIVE state (among other problems). Add code to
qib_ud_loopback() to correctly initialize the pkey_index.

INTERNAL: PR 125363 - Enabling SmaSpoofingCheck causes invalid P_key
INTERNAL: PR 125363 errors, port bounce leaves port in Init state

Reviewed-by: Ira Weiny <ira.weiny@intel.com>
Signed-off-by: Arthur Kepner <arthur.kepner@intel.com>

Change-Id: Ib0bca5ad8f2ecde19ca2d026a02725a72bced286