IB/hfi: Remove skip_nn_check workaround
This removes the workaround option skip_nn_check as all systems
have PRR v137 or higher.

Add a simulator workaround for setting NeighborNormal.  The simulator
does not pass SMA messages which set NeighborNormal, so just set it
when moving to Active.

INTERNAL: PR 124812 - PRR, WFR not setting NeighborNormal when neighbor
INTERNAL: gets to Init or Armed and they are going active while
INTERNAL: NeighborNormal is zero.

Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Dean Luick <dean.luick@intel.com>