btt: initial implementation
BTT stands for Block Translation Table, and is a way to
provide power fail sector atomicity semantics for block
devices that have the ability to perform byte granularity
IO. It introduces a new block_device_operations entry point,
rw_bytes. A driver intending to use the BTT must provide
this entry point.

The BTT works as a stacked blocked device, and reserves a chunk
of space from the backing device for its accounting metadata.

Signed-off-by: Vishal Verma <vishal.l.verma@linux.intel.com>
Signed-off-by: Dan Williams <dan.j.williams@intel.com>