IB/hfi: Use new link width definitions
Storm Lake is changing its link width defines.   All non-supported
sizes are dropped and the defines are changing.  Change the code
to stop using any IB link width define and any non-supported size.

There is a coordination issue with the external header that
contains the link width defines.  If the new defines are not found,
override the old defines and define the new ones locally.  When
the new defines are in the proper header the override will be
removed.

INTERNAL: PR 125707 - Update WFR to use new STL Link Width definitions

Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Dean Luick <dean.luick@intel.com>