IB/hfi: correct two AHG flaws
The logic in submit_tx that or's in the generation number needs
to be avoided for the up to 2 descriptors that contain edits.

This patch resolves that issue by computing the number of
descriptors (skip) that follow the first and not adding the
generation number for those descriptors.

To faciliate the patch, the first descriptor is pulled out of the
loop to avoid a special case test when downcounting the skip
count.

There is a second issue where not all the 32b AHG descriptors
were copied into the 16B SDMA descriptors.  This flaw is
also corrected.

Endian issues are also resolved in this patch.

INTERNAL: PR 125720 - AHG mid-layer flaws found during PSM/AHG debug

Change-Id: I24ca119e5450be40ade638e9b705cd61daabf7b1
Reviewed-by: Mitko Haralanov <mitko.haralanov@intel.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>