IB/hfi:Zero out reserved capture metadata fields
When metadata was introduced for the capture capability of the driver there was
some space left in the structure for future use. This "reserved" space was not
explicitly zeroed and as such may contain garbage data. This patch remedies that
by zeroing the reseved fields.

INTERNAL: PR 125395 - Snoop and capture Header Reserved Fields are non-zero

Reviewed-by: Dean Luick <dean.luick@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Change-Id: Ibba38b0265d7e8c5a8d16ff6bd4b190f67244f37