nd_dsm: Initial version of manual _DSM IOCTL path
Implements an IOCTL path that manually run vendor specific _DSM commands
by directly interacting with the Falcon Valley mailbox registers, as
opposed to using the ACPI _DSM method via the BIOS.  This is meant to be
a stopgap while _DSM support is being implemented in the BIOS, and
should eventually be removed from our driver.

A lot of the code in this patch series is copied from the CRBD driver,
so it will definitely need to be cleaned up and sanitized (no mention of
Falcon Valley, fnv, Crystal Ridge, CR, CRBD, etc) if we ever want to
give this to anyone.

Also, at present this code only works for the single AEP NVDIMM
configuration in simics, and is hard coded so that it has no dependency
on the NFIT.  If/when the NFIT becomes consistently available we can
make this code dynamic, and we can also consider supporting multiple
NVDIMMs if necessary.

Signed-off-by: Ross Zwisler <ross.zwisler@linux.intel.com>