IB/hfi: Send idle flit for active and armed in snoop ioctl.
Snoop IOCTL handler needs to be able to trigger the idle flit to be sent in
order to set neighbor normal on the other side. This is normally done by the SMA
but when snoop is active that won't happen.

INTERNAL: PR 125794 - WFR Driver ioctl setlinkstateextra doesn't send
INTERNAL: NeighborNormal idle flit.

Reviewed-by: Dean Luick <dean.luick@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Change-Id: I69a3d5530546bb0f88ecdaee750856591d0c0a42