IB/hfi: place active CCA state in an RCU protected structure
When CCA is invoked the active CCA configuration data may need to be
read in a fast path, so place that data in an RCU-protected struct,
cc_state.

Note that the per-SL 'congestion setting' data, and the per-port
'congestion control table' are always read from within cc_state in the
relevant SMA 'get'. But the 'congestion setting' data, and 'congestion
control table' are initially written to 'shadow' data structures in the
corresponding 'set' methods. The cc_state data is copied from the
shadow structures as one of the final steps in a 'set' of the per-port
congestion control table.

In addition to keeping CCA data in an RCU-protected structure, this
patch also:
  - removes some unused structure definitions from mad.h
  - makes the 'congestion setting', and 'congestion control table'
    shadow arrays statically allocated members of struct qib_pportdata

INTERNAL: PR 125617 - CCA support for verbs

Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Arthur Kepner <arthur.kepner@intel.com>

Change-Id: I9ee41c0a2cfe08f959167c7bb7e60869eea0049a